pyproject.toml
[project]
name = "whispper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bullmq>=2.13.1",
    "flask>=3.1.0",
    "librosa>=0.11.0",
    "openai-whisper>=20240930",
    "python-dotenv>=1.1.0",
    "quart>=0.20.0",
    "torch>=2.6.0",
    "werkzeug>=3.1.3",
]
